주문목록 보기 삭제 위시리스트 보기 및 위시리스트에서 삭제

+++ AnnaSemiProject/src/main/java/member/controller_DH/RemoveItemAction.java
@@ -4,8 +4,12 @@ import java.util.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import org.json.JSONObject;
 
 import common.controller.AbstractController;
+import member.model.MemberVO;
 import product.model_lsh.*;
 
 public class RemoveItemAction extends AbstractController {
@@ -32,28 +36,36 @@ public class RemoveItemAction extends AbstractController {
 		else {
 			// "POST"방식일 경우
 			
-			String productnum = request.getParameter("removeItemNum");
+			String productnum = request.getParameter("productnum");
 			String userid = request.getParameter("userid");
 			
-			InterProductDAO pdao = new ProductDAO();
+			HttpSession session = request.getSession();
+			MemberVO loginuser = (MemberVO)session.getAttribute("loginuser");
 			
-			Map<String, String> paraMap = new HashMap<>();
 			
-			paraMap.put("productnum", productnum);
-			paraMap.put("userid", userid);
-			
-			int n = pdao.removeWishProduct_DH(paraMap);
-			
-			if(n == 1) {
-				message = "삭제가 완료되었습니다.";
-				loc = "javascript:history.back()";
+			if(loginuser.getUserid().equals(userid)) {
+				InterProductDAO pdao = new ProductDAO();
 				
-				request.setAttribute("message", message);
-				request.setAttribute("loc", loc);
+				Map<String, String> paraMap = new HashMap<>();
+				
+				paraMap.put("productnum", productnum);
+				paraMap.put("userid", userid);
+				
+				int n = pdao.removeWishProduct_DH(paraMap);
+				
+				JSONObject jsobj = new JSONObject();
+				jsobj.put("n", n);
+				
+				String json = jsobj.toString();
+				
+				request.setAttribute("json", json);
 				
 				super.setRedirect(false);
-				super.setViewPage("/WEB-INF/msg.jsp");
+				super.setViewPage("/WEB-INF/jsonview.jsp");
 			}
+			
+			
+		
 			
 			
 		}
